Campaigns: Introduce mapillary filter; Rework naming to "todoKey(s)"
We rename projectKey to todoKey because we call them todos and todos_lines (DB table).

We split campaign id and todoKey so we can have multiple campaigns with the same todoKey. That breaks our premise to have a 1on1 relation between the two. But as long as we have the  mapillary_coverage data added after osm2pgsql runs, this is the cleaner way.

We add a new category for with to promote counting stones.