Notes: Fix missplaced note pin when using the feature links
Somewhere while passing the lat-lng of the current object we applied rounding based on zoom 12 which removed too much precision for the pin. No idea why that rounding happened only while submitting right before submitting and not when we first set the value … but increasing the initial value resolved the issue.