Parking: Rename _parking_parkings<Number>_<Desc> table
Encode the step / process into the table name with 1,2,3
Add a description to the step with "road" for the source table, "cut" for data after it was "cutout", "merged" after the merge (and capacity calc) step.
We don't have final data that is non-prefixed and follows our vector tile table convention, yet.