Remove updated_age from processing and style
This property was expensive in terms of storage. It made our diffing show unhelpful data. And we actually don't need it, once we handle the one place where we handle the age of the objects differently.

This also cleans up old code about other *_age properties that where unused for a while now.