Handle construction/temporary ways
Ways in construction are no longer skipped but transformed into their future (previous) infrastructure (roads, path, bikelanes).
The new key livecycle=* is added to indicate this change.
The Inspector shows this "Status".
The key temporary=yes (Alternative way during construction) is transformed to livecycle as well.

Some refactoring was needed to get this implemented in a cleanish way.