Roads, Bikelanes: Add internal _width_effective tag
Ping https://github.com/FixMyBerlin/private-issues/issues/2939

We want to rework the way we present width data in TILDA. But this will take some time. For now we just pass this key for internal use. Its not part of the CHANGELOG since we consider it a property that might change any time.